Extensive rewrite of function stringification
Node.js 10 changed how the toString method on functions works. The
resulting string is much less ambiguous than what the method produces on
older versions, but also requires a more robust parser to interpret
correctly. Stripping off computed property prefixes necessitates the
ability to find the matching ']' to a given '[', and doing that
correctly means taking into account all the ways that comments, template
strings, and regular expression literals can fool a naive
implementation. A full JavaScript parser would be many times the amount
of code added here; this "parser-lite" implementation cuts many corners
but is possibly bug-free on the most recent version of Node.

The toString implementation of previous versions of Node continues to be
supported. On Node.js 4, this logic is also possibly bug-free. However,
a behavior change that I believe started in Node.js 6 means that
versions of Node.js greater than 4 and less than 10 can toString a small
subset of functions in such a way that it is impossible to unambiguously
stringify them in all circumstances. Still, we do the best we can. Tests
that are conditionally skipped on the affected Node.js versions
demonstrate the problem.